Add files via upload
The raspberry has been programmed to measure the body temperature. The body temperature is digitally displayed on a display controlled by the same raspberry that transmits the temperatures simultaneous to the web server using Wi-Fi technology. In the case of a high temperature a message indicates to the person to take a pill. The significance of the body temperature monitor is that it provides an inexpensive and accurate means of measuring one's health condition at his/her convenience.